% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_macarons.R
\name{theme_macarons}
\alias{theme_macarons}
\title{Génération d'un thème macarons ggplot2}
\usage{
theme_macarons(my_font)
}
\arguments{
\item{my_font}{police google font à utiliser}
}
\value{

}
\description{
Génération d'un thème macarons ggplot2
}
